fix mirror dates
